requirements.txt
PyQt5==5.15.9
PyQt5-sip==12.15.0
tensorflow==2.18.0
mysql-connector-python==8.1.0
Pillow==10.0.0
numpy==1.26.4

diffusers==0.14.0
transformers==4.30.0



